Move PouchDB Error definitions to a separate package to avoid inclusion in production builds.
