+++ driver/pouchdb/bindings/poucherr/poucherr.go
@@ -1,0 +1,4 @@
+// Package poucherr exists only for the purpose of testing the PouchDB binding's
+// handling of PouchDB-specific error messages. This Go package is empty, the
+// real logic is in poucherr.inc.js.
+package poucherr
